+++ src/data/rules/conjunct-rules.js
@@ -1,0 +1,6 @@
+export const conjunctRules = {
+  kkh: "ক্ষ",
+  ny: "ন্য",
+  gn: "জ্ঞ"
+};
+