+++ src/data/rules/kar-rules.js
@@ -1,0 +1,12 @@
+export const karNames = {
+  "া": "aa-kar",
+  "ি": "i-kar",
+  "ী": "ii-kar",
+  "ু": "u-kar",
+  "ূ": "uu-kar",
+  "ে": "e-kar",
+  "ৈ": "oi-kar",
+  "ো": "o-kar",
+  "ৌ": "ou-kar"
+};
+